Updated openstack/openstack Project: openstack/nova  b9ce1240e4e510f7bbc7f49ee44b866ed13bda3a
Fix the duplicated image params in a test

In _test_create, there is a parameter "no_image" to disable "image_ref"
in a request body of "create a server" API. The way of disabling is
setting '' as "image_ref".
So the unit tests using _test_create pass the duplicated images:
the parameter image_ref '' and the parameter block_device_mapping

This patch fixes this duplicated image parameters.
This problem was detected through the API validation works and
Ic17d2227909c9180ca86c9d578dad4c73ec57dfc needs this fixing.

Change-Id: I36ffc87309afadec7da3fa19e01ea8f00f0f9f8d