Updated openstack/openstack Project: openstack/nova  8448ac6118660be6549a0dda33a1f5f01ccabe2d
Remove bdms from run_instance RPC conductor call

After the move to build_and_run_instance in change
Iefab71047996b7cc08107794d5bc628c11680a70 we get the block device
mapping out from the DB on the conductor side. We need to do this since
there is not enough data in the RPC call from the API (we don't have DB
id's so compute can't use that data to update block devices in the boot
process).

Based on this - we should stop sending block devices to the conductor.

Change-Id: Ic7a5da3f0d58093c3b249025b9fcd36a1321f2b3