Updated openstack/openstack Project: openstack/requirements  2a5e0b1d47d6c94f2c66156ca5991a0cd73e5a57
Pin httpretty's precise version.

The latest release, 0.8.1, has completely broken installation, so for now block
it: https://github.com/gabrielfalcao/HTTPretty/pull/175 Can be unpinned once
they issue a release with a fix.

The 0.8.2 release is also broken, so exclude both of those releases.

Sounds like the fix is in 0.8.3 which is on pypi now.

Closes-Bug: #1332266
Change-Id: I42376a11cbba58184f3dd99245e0b39136d88ff4