Updated openstack/openstack Project: openstack/taskflow  14d0ceef3538f526a54845ab31f850bc0b801d67
Sync our version of the interprocess lock

The oslo version gained a exists function and
merged in the commented code we had so we can
go ahead and merge that into our source tree
and remove the comment (since it's not useful
anymore).

The update is from lockutils.py, oslo incubator
commit hash e3b52f20367e18.

Change-Id: I858a4412d03c48c4ef1c2506edf6334c66ded3f0