Updated openstack/openstack Project: openstack/taskflow  e328f133b9d52f08b8697b33bee0c8285f6eb5ab
Make intentions a tuple (to denote immutability)

Since the atom intentions are not meant to be mutable
we should make sure that we create the object as a type
which is not mutable. This helps with those looking over
the code to understand its desired usage which is not to
be a mutable collection.

Change-Id: I84948faf2e6bd8f4b4a9b27e390c9a03a14efa4b