Updated openstack/openstack Project: openstack/python-heatclient  a1168900dc3c66dc3d2ea9e2081dc130635dea8f
Do not set up logging handler in http

A logging handler should not be set up in the http module as this
can cause unintended debug logging of all requests when
python-heatclient is used as a library. This is causing log bloat
in tempest gate runs.

This change removes the logging handler from the http module.
This does not affect the --debug option of the heat cli.

Change-Id: I634dc3bb96f3880df322bd627c9075ebac13add8
Closes-Bug: #1330492