Updated openstack/openstack Project: openstack/horizon  a25fe2848f21eae3aa35a360cd804e2f4fcd4421
Upgrade jquery to version 1.10.2 with jquery-migrate

We are using an ancient version of jquery, and before unbundling it,
it would be nice to upgrade it. Unfortunately, some other libraries,
such as bootstrap.js, use some of the deprecated functions. That's
why we will need to have jquery-migrate.js until we upgrade those too.

This patch upgrades jquery.js to version 1.10.2, adds jquery-migrate.js
version 1.2.1, and fixes some of the deprecated stuff in the horizon's
own libraries.

Change-Id: Ifcb0a7132f365816794c180b5979f57ab23b513b