Updated openstack/openstack Project: openstack/cinder  c3240b96e84adda494ce08ce05f8ea81d1f158e4
Fix order dependency of admin metadata

Some test cases in VolumeTestCase depends on the order of admin metadata
returned from db.volume_get() call, which causes jenkins test failure in
recent reviews.  This change extracts the key/value we are intertested in
and compare them with AssertDictMatch() to remove the order dependency.

Change-Id: I2719c025436304cabab54f19a6ce314842d8785b
Closes-bug: # 1332388