Updated openstack/openstack Project: openstack/horizon  4a96f2e2a2b087508c369e0bea891a29829a673d
Fix delete button interface in network topology view

In the Network Topology panel, there is a pop-up for the router which
includes a 'Delete Interface' button for each port on the router.
Clicking this button does not actually delete the port.

The fix add a tab_group render in iframe.html, since the view
can actually be a TabView.

Change-Id: I274f5165ca31841cb2a2f89ab196a740c2644df6
Closes-Bug: #1282811