Updated openstack/openstack Project: openstack/neutron  4482f4669070fb86b87f747bba678e1950650216
Check port value when creating firewall rule with icmp protocol

If a firewall rule was created with setting protocol as icmp
and using source/destination port no error was thrown, even though
source/destination port parameters are not used by icmp.

This patch adds a validation method that checks passed parameters
and throws an exception if they are incompatible.

Change-Id: I90a765856896395fcb6e9ddbd888b7bd80480674
Closes-Bug: 1327955