Updated openstack/openstack Project: openstack/neutron  b56b738a16a14acec102ad629ef0adcc52d6fa7a
linuxbridge-agent: process port updates in the main loop

This patch changes the way the linuxbridge agent processes port update
notifications. It does the same for the lb-agent as was done for the
ovs-agent in I219c6bdf63b0b5e945b655677f9e28fa591f03cd.
Now the RPC call just adds the updated tap device name to a set of
updated devices, and the actual processing is done in the main RPC loop.
This should solve the problems were port_update RPCs were competing
with the main loop/greenthread and with each other to do the same task
- like adding a tap or interface to a bridge - which lead to races and
was inefficient. Now repeated port_update notifications received within
a single main loop iteration will be coalesced and processed only once.

Closes-Bug: 1256950
Change-Id: I7fd48542f12b39ffc1346d1a6c9a387ecda6d812