Updated openstack/openstack Project: openstack/taskflow  630364b4c13ae6f1aa6cdf9a2219239593e639dd
Improve WBE testing coverage

Test more of the message pump process and verify
that messages are received and sent correctly by
using a latch type and waiting for the desired
number of messages to be triggered.

Test the full server and executor pipeline by
setting up threads and a in-memory queue/exchange
that is used by these threads for task execution.

Change-Id: I62296f12eee6fa00559e84068ec5ee2a6d4bc0dc