Updated openstack/openstack Project: openstack-dev/pbr  b07a50bfddda1cbb10faa64dc096ca9d855c483c
Restore Monkeypatched Distribution Instance

pbr is monkeypatching the Distribution intstance in setuptools, which is
OK I guess, but then it never puts it back how it left it.  As a
consequence, whenever pbr installs before another project in a setup.py,
it breaks setuptools for everything that comes after it.  This patch
will restore the Distribution instance.

Closes-Bug: 1324784

Change-Id: I20fbf255c634685e82f7b11987d2725de8280b9d