Updated openstack/openstack Project: openstack/oslo-incubator  38ee366d7ac501e4cc6a430aec28751d9bbf8661
Enable test_strutils for Python 3

common.strutils is compatible with Python 3, and thus test_strutils can
be enabled. Also in test_strutils there was implicit  assumption, that
sys.stdin.encoding is either utf-8 or ascii, so symbols such as \x80 are
undecodable. If it happens not to be true, tests are failing.

Change-Id: I0e2decd8d4ad58c999b8ecbab2c5e921dc8b1df8