Updated openstack/openstack Project: openstack/nova  e3ddd7c354c2eebebecbb6edab17a0551b3ddb0e
Catch permission denied exception when update host

When an existing non-openstack instance in the compute node contains
disk file which can not be accessed by nova non-root user, resources can
not be updated successfully when periodic task is running. Also the
nova-compute service could not be restared again. We should catch such
exception.

Change-Id: I75195417e123f135252470cf1322843830556b0a
Closes-Bug: 1322467