Updated openstack/openstack Project: openstack/nova  b37c4ad695e5fcbd59ed02d0606fc7080e951c97
Allow evacuate from vm_state=Error

We currently allow reboot/rebuild/rescue for an instance in the Error state.
This commit allows "evacuate" as well, since it is essentially a "rebuild"
on a different compute node.

This is useful in a number of cases, in particular if an initial evacuation
attempt fails.

Change-Id: I3f513eb738c91fe71767308f57251629639efd6a
Closes-Bug: 1298061