Updated openstack/openstack Project: openstack/cinder  e20ef8a8bb56297892b05d2a5bd728f503dd637e
Fixed data copy issue of volume/driver.py

I found data copy issue on __init__ of ISERDriver class.
Some settings for iser should be copied to iscsi's.
But iser_port value was copied not to iscsi_port but to iser_port.
This patch will fix this issue as to be copied to iscsi_port.

Change-Id: I7621e690396049aabf3ced331e633158f45067f4